backend/trekapi/requirements.txt
asgiref==3.8.1
Django==4.2.18
django-cors-headers==4.7.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
mysqlclient==2.2.7
PyJWT==2.10.1
sqlparse==0.5.3
typing_extensions==4.12.2
